requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.70.0
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.9
Jinja2==3.1.4
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
urllib3==2.3.0
uv==0.6.3
Werkzeug==3.1.3
wheel==0.45.1
yacs==0.1.8
sentencepiece-0.2.0
bitsandbytes-0.45.5